create gateway file
